requirements.txt
# YouTube Auto Factory — Dependencies
# Install: pip install -r requirements.txt

playwright==1.49.1
Pillow>=10.0.0
moviepy>=1.0.3
schedule>=1.2.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
ollama>=0.3.0
python-dotenv>=1.0.0
google-genai>=1.0.0
